Move functions to utils files for cleaner init command
